requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
diff-match-patch==20241021
dj-database-url==3.0.1
dj-redis-panel==0.6.0
Django==5.2.8
django-celery-beat==2.8.1
django-import-export==4.3.14
django-jazzmin==3.0.1
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.1
dotenv==0.9.9
frozenlist==1.8.0
gunicorn==23.0.0
idna==3.11
kombu==5.5.4
multidict==6.7.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.0.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tablib==3.9.0
twilio==9.8.6
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
yarl==1.22.0
